+++ CHANGELOG.md
@@ -25,7 +25,7 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
     - `Py42TrustedActivityInvalidCharacterError`
     - `Py42TrustedActivityIdNotFound`
 
-- New custom py42 `409` error wrapper class `Py42ConflictError`
+- New custom `HTTP 409 error` wrapper class `Py42ConflictError`
 
 ### Fixed